fix: fixed reset button for flags

+++ backend/src/routes/api/navigate/group.js
@@ -40,6 +40,20 @@ const deleteAllNotes = async (groupData) => {
   return true;
 };
 
+const deleteAllFlags = async (groupData) => {
+  const groupId = groupData._id;
+  const group = await Group.findById(groupId).lean();
+  if (!group) {
+    throw new HttpError("Group not found", STATUS.NOT_FOUND);
+  }
+
+  await Group.updateOne({ _id: groupId }, { $set: { currentFlags: [] } }).exec();
+  // if (res.nModified !== 1) {
+  //   throw new HttpError("Failed to delete notes", STATUS.INTERNAL_SERVER_ERROR);
+  // }
+  return true;
+};
+
 export const getScenarioFirstScene = async (scenarioId) => {
   const scenario = await Scenario.findById(scenarioId, {
     scenes: { $slice: 1 },
@@ -189,6 +203,9 @@ export const groupNavigate = async (req) => {
 export const groupReset = async (req) => {
   const { uid, currentScene } = req.body;
   const group = await getGroupByIdAndUser(req.params.groupId, uid);
+  if (!group) {
+    throw new HttpError("Group not found", STATUS.NOT_FOUND);
+  }
   const { role } = group.users[0];
 
   if (!(await deleteAllNotes(group)))
@@ -198,15 +215,20 @@ export const groupReset = async (req) => {
     throw new HttpError("Scene mismatch has occured", STATUS.CONFLICT);
 
   const scene = await getSceneConsideringRole(currentScene, role);
+  if (!scene || !scene.components) {
+    console.error("Invalid scene or missing components:", scene);
+    throw new HttpError("Scene not found or invalid", STATUS.NOT_FOUND);
+  }
+
   const hasReset = scene.components.some((c) => c.type === "RESET_BUTTON");
   if (!hasReset) throw new HttpError("Invalid reset", STATUS.FORBIDDEN);
 
-  await Group.findOneAndUpdate(
-    { _id: group._id },
-    { $set: { path: [], currentFlags: [] } },
-    { new: true }
-  );
 
+  console.log(group)
+
+  if (!(await deleteAllFlags(group)))
+    throw new HttpError("Failed to delete flags", STATUS.INTERNAL_SERVER_ERROR);
+  
   return { status: STATUS.OK };
 };
 
@@ -220,6 +242,8 @@ export const groupGetResources = async (req) => {
 
   const flags = group.currentFlags || [];
   const resources = [];
+
+  console.log(flags)
 
   // Fetch all resources from the database
   const allResources = await Resource.find({});
